Fixed compile errors
"auxilary" -> "auxiliary"
